port another variable, seemed to fill in "white" accidentally
