+++ enclone_vars/src/vars
@@ -2749,3 +2749,50 @@ avail:    public
 notes:
 code:     exact: format!("{}", ex.share[mid].j_stop - ex.share[mid].v_start)
 ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+name:     white
+inputs:   cvar_vdj
+limits:
+class:    cvar
+level:    exact
+val:      float
+doc:      TBD
+brief:    percent of sequences implicated in whitelist expansion
+page:     UNDOCUMENTED
+avail:    private
+notes:
+code:     let mut bch = vec![Vec::<(usize, String, usize, usize)>::new(); 2];
+          for l in 0..ex.clones.len() {
+              let li = ex.clones[l][0].dataset_index;
+              let bc = &ex.clones[l][0].barcode;
+              let mut numi = 0;
+              for j in 0..ex.clones[l].len() {
+                  numi += ex.clones[l][j].umi_count;
+              }
+              bch[0].push((li, bc[0..8].to_string(), numi, l));
+              bch[1].push((li, bc[8..16].to_string(), numi, l));
+          }
+          let mut junk = 0;
+          let mut bad = vec![false; ex.clones.len()];
+          for l in 0..2 {
+              bch[l].sort();
+              let mut m = 0;
+              while m < bch[l].len() {
+                  let n = next_diff12_4(&bch[l], m as i32) as usize;
+                  for u1 in m..n {
+                      for u2 in m..n {
+                          if bch[l][u1].2 >= 10 * bch[l][u2].2 {
+                              bad[bch[l][u2].3] = true;
+                          }
+                      }
+                  }
+                  m = n;
+              }
+          }
+          for u in 0..bad.len() {
+              if bad[u] {
+                  junk += 1;
+              }
+          }
+          let junk_rate = percent_ratio(junk, ex.clones.len());
+          exact: format!("{:.1}", junk_rate)
+━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
